Fix light tree + linking crash with shared subtrees
The assumption that the left child node follows the parent immediately is
not true in this case. For now just store both the left and right child index
as there is already padding anyway.

Later this could be optimized by storing the left and right child next to
each other, and then ensuring both are always specialized together.