Fix excluded light linking behavior
The light intersection was always using the default receiver set
because the receiver_forward was always set to OBJECT_NONE.

This also fixes the case when there is a scene with a single object,
multiple lights lights linked to it, and one default light: the
render result now matches with the main branch.

This breaks the regression test.
The brighter result in the `light link simple` is expected, but the
`light link simple tree` seems to be too bright.

Pull Request: https://projects.blender.org/blender/blender/pulls/107973