Light linking: Reduce noise for shadow linked distant lights
If there is a default blocker at the path to the distant light
the code did not ignore distant lights.

The blockers from the default set were already leading to the
other light types from being ignored because of the assignment
of the ray->tmax.

Pull Request: https://projects.blender.org/blender/blender/pulls/108022